+++ chain/src/indexer/backfiller/mod.rs
@@ -2,7 +2,8 @@
 //!
 //! The backfiller path is split into:
 //! - [`state`], which owns the shared upload state and the application-side
-//!   [`Recorder`] that persists finalized block digests; and
+//!   data structures used by the uploader actors;
+//! - [`recorder`], which persists finalized block digests; and
 //! - [`drainer`], which owns the background retry loop that drains those queue
 //!   rows and uploads blocks.
 //!
@@ -10,7 +11,9 @@
 //! [`SharedUploadState`].
 
 mod drainer;
+mod recorder;
 mod state;
 
 pub use drainer::Drainer;
-pub use state::{FinalizedEntry, Recorder, SharedUploadState, UploadDecision, UploadState};
+pub use recorder::Recorder;
+pub use state::{FinalizedEntry, SharedUploadState, UploadDecision, UploadState};
